requirements.txt
contourpy==1.3.3
cycler==0.12.1
Faker==37.5.3
fonttools==4.59.0
greenlet==3.2.4
joblib==1.5.1
kiwisolver==1.4.8
lightgbm==4.6.0
matplotlib==3.10.5
networkx==3.5
numpy==2.3.2
packaging==25.0
pandas==2.3.1
pillow==11.3.0
psycopg2-binary==2.9.10
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
scikit-learn==1.7.1
scipy==1.16.1
seaborn==0.13.2
six==1.17.0
SQLAlchemy==2.0.42
threadpoolctl==3.6.0
typing_extensions==4.14.1
tzdata==2025.2
